:recycle: Refactor models to individual files
